// frontend/src/backend.d.ts
export interface LocalCivicBody {
    photoPath?: string;
    representativeName: string;
    bodyName: string;
    bodyType: string;
}
export interface FileReference {
    hash: string;
    path: string;
}
export interface Directory {
    states: Array<State>;
    unionTerritories: Array<State>;
    administrativeUnits: Array<AdministrativeUnit>;
    primeMinister?: Representative;
}
export interface Representative {
    politicalParty?: string;
    photoPath: string;
    name: string;
    lastUpdated: bigint;
    email: string;
    twitterHandle: string;
    remarks: string;
}
// ...
export interface Report {
    id: string;
    status: string;
    issueType: string;
    photoPath: string;
    username?: string;
    resolvedByVolunteer: boolean;
    mpData?: Representative;
    reporterName?: string;
    customAddress?: string;
    cmPhotoPath?: string;
    mlaDesignation: string;
    submittedByVolunteer: boolean;
    state: string;
    cmData?: Representative;
    cmName?: string;
    address: string;
    notes?: string;
    timestamp: bigint;
    pmData?: Representative;
    localCivicBody?: LocalCivicBody;
    pmName?: string;
    location: {
        latitude: number;
        longitude: number;
    };
    mlaPhotoPath?: string;
    proofPhotoPath?: string;
    pmPhotoPath?: string;
    mlaName?: string;
    coordinates: string;
    completionNotes?: string;
}